Fixed AddContactCommand being push into undo Stack when adding failed & moved VFD action to navigateReceiver
