+++ src/contactmanagementsoftware/MUI.java
@@ -131,11 +131,11 @@ public class MUI extends JFrame {
         addContactCommand = new AddContactCommand(contactReceiver);
         deleteCommand = new DeleteCommand(contactReceiver);
         searchCommand = new SearchCommand(contactReceiver);
-        vfdCommand = new VFDCommand(contactReceiver);
         readFromFileCommand = new ReadFromFileCommand(contactReceiver);
         saveCommand = new SaveCommand(contactReceiver);
         addCommand = new AddCommand(navigateReceiver);
         editCommand = new EditCommand(navigateReceiver);
+        vfdCommand = new VFDCommand(navigateReceiver);
         backToMainMenuCommand = new BackToMainMenuCommand(navigateReceiver);
         cancelCommand = new CancelCommand(navigateReceiver);
         exitCommand = new ExitCommand(navigateReceiver);
@@ -598,84 +598,6 @@ public class MUI extends JFrame {
             
         }
     }//GEN-LAST:event_jList1ValueChanged
-
-    public void searchNameAndDisplay(){
-        String s = "<html> <b>Search results:</b><br>Found!<br><br>Acquaintance Details: <br>";
-        int j = 0;
-        Expression searchExpression = new Parser().constructParser(str);
-        for(int i = 0; i < a.get(0).size(); i++){
-            //if(a.get(0).get(i).getName().matches(str)){
-            if(searchExpression.interpret(a.get(0).get(i).getName())){
-                j++;
-                PersonalFriends perF = (PersonalFriends)a.get(0).get(i);
-                if(j==1){
-                    s = s.concat("<br>I. Personal Friends<br>");
-                }
-                s = s.concat(j + ". Name: " + perF.getName() + "<br>");
-                s = s.concat("Mobile No: " + perF.getMobileNo() + "<br>");
-                s = s.concat("Email: " + perF.getEmail() + "<br>");
-                s = s.concat("Specific events: " + perF.getEvents() + "<br>");
-                s = s.concat("First Acquaintance context: " + perF.getAContext() + "<br>");
-                s = s.concat("First Acquaintance date: " + perF.getADate() + "<br>");
-            }
-        }
-        j = 0;
-        for(int i = 0; i < a.get(1).size(); i++){
-            //if(a.get(1).get(i).getName().matches(str)){
-            if(searchExpression.interpret(a.get(1).get(i).getName())){
-                j++;
-                Relatives rel = (Relatives)a.get(1).get(i);
-                if(j==1){
-                    s = s.concat("<br>II. Relatives<br>");
-                }
-                s = s.concat(j + ". Name: " + rel.getName() + "<br>");
-                s = s.concat("Mobile No: " + rel.getMobileNo() + "<br>");
-                s = s.concat("Email: " + rel.getEmail() + "<br>");
-                s = s.concat("Relatives Birthday: " + rel.getBDate() + "<br>");
-                s = s.concat("Last met date: " + rel.getLDate() + "<br>");
-            }
-        }
-        j = 0;
-        for(int i = 0; i < a.get(2).size(); i++){
-            //if(a.get(2).get(i).getName().matches(str)){
-            if(searchExpression.interpret(a.get(2).get(i).getName())){
-                j++;
-                ProfessionalFriends proF = (ProfessionalFriends)a.get(2).get(i);
-                if(j==1){
-                    s = s.concat("<br>III. Professional Friends<br>");
-                }
-                s = s.concat(j + ". Name: " + proF.getName() + "<br>");
-                s = s.concat("Mobile No: " + proF.getMobileNo() + "<br>");
-                s = s.concat("Email: " + proF.getEmail() + "<br>");
-                s = s.concat("Common Interests: " + proF.getCommonInterests() + "<br>");
-            }
-        }
-        j = 0;
-        for(int i = 0; i < a.get(3).size(); i++){
-            //if(a.get(3).get(i).getName().matches(str)){
-            if(searchExpression.interpret(a.get(3).get(i).getName())){
-                j++;
-                CasualAcquaintances ca = (CasualAcquaintances)a.get(3).get(i);
-                if(j==1){
-                    s = s.concat("<br>IV. Casual Acquaintances<br>");
-                }
-                s = s.concat(j + ". Name: " + ca.getName() + "<br>");
-                s = s.concat("Mobile No: " + ca.getMobileNo() + "<br>");
-                s = s.concat("Email: " + ca.getEmail() + "<br>");
-                s = s.concat("First met location & time: " + ca.getWhenWhere() + "<br>");
-                s = s.concat("First met circumstances: " + ca.getCircumstances() + "<br>");
-                s = s.concat("Other useful information: " + ca.getOtherInfo() + "<br>");
-            }
-        }
-        if(s.matches("<html> <b>Search results:</b><br>Found!<br><br>Acquaintance Details: <br>")){
-            s  = "<html>No result found</html>";
-        }
-        else{
-            s = s.concat("</html>");
-        }
-        details.setText(s);
-    }
-
 
     public boolean MobileNoChecker(String str){
         int x;
