+++ src/contactmanagementsoftware/command/Command.java
@@ -4,7 +4,7 @@ import java.awt.event.ActionEvent;
 
 public interface Command {
 
-    public void execute(ActionEvent evt);
+    public boolean execute(ActionEvent evt);
 
     public void undo(ActionEvent evt);
 }
