+++ src/contactmanagementsoftware/command/ContactReceiver.java
@@ -1,6 +1,8 @@
 package contactmanagementsoftware.command;
 
 import contactmanagementsoftware.*;
+import contactmanagementsoftware.Interpreter.Expression;
+import contactmanagementsoftware.Interpreter.Parser;
 import contactmanagementsoftware.singleton.Logger;
 import contactmanagementsoftware.singleton.LoggerSingleton;
 
@@ -11,18 +13,22 @@ import java.util.ArrayList;
 
 public class ContactReceiver {
     static MUI mui;
+    String str;
     JPanel jPanel1, jPanel2, jPanel3;
-    int selectedContactTypeIndex = -1, selectedContactIndex = -1;
+    JTextPane details;
     ArrayList<ArrayList<Acquaintances>> a, temp;
     Acquaintances acquaintance;
     ArrayList<Integer> uploadedFileIndexes;
+    int selectedContactTypeIndex = -1, selectedContactIndex = -1;
     private Logger logger = LoggerSingleton.getInstance();
 
     public ContactReceiver(){
         mui = mui.getInstance();
+        str = mui.getStr();
         jPanel1 = mui.getjPanel1();
         jPanel2 = mui.getjPanel2();
         jPanel3 = mui.getjPanel3();
+        details = mui.getDetails();
         uploadedFileIndexes = new ArrayList<>();
         a = mui.getAllAcquantanceList();
     }
@@ -221,26 +227,6 @@ public class ContactReceiver {
         mui.setUpTableData();
     }
 
-    public void viewFullDetail(){
-
-        if(mui.getSelectedContactTypeIndex() < 0){
-            JOptionPane.showMessageDialog(mui, "Select a category!");
-            return;
-        }
-        
-        if(mui.getSelectedContactIndex() < 0){
-            JOptionPane.showMessageDialog(mui, "Select an entry!");
-            return;
-        }
-
-        mui.setIsAddContact(false);
-        jPanel1.setVisible(false);
-        jPanel3.setVisible(true);
-        mui.setIsDisplayOnly(true);
-        mui.contactDetailsPanelSetterSetUI();
-
-    }
-
     public void searchContact() {
         JTextPane details = mui.getDetails();
 
@@ -258,7 +244,7 @@ public class ContactReceiver {
         jPanel2.setVisible(true);
         mui.setStr(s);
         details.setContentType( "text/html" );
-        mui.searchNameAndDisplay();
+        searchNameAndDisplay();
     }
 
     public boolean uploadContacts(){
@@ -330,6 +316,83 @@ public class ContactReceiver {
             return;
         }
         JOptionPane.showMessageDialog(mui, s + " saved successfully");
+    }
+
+    public void searchNameAndDisplay(){
+        String s = "<html> <b>Search results:</b><br>Found!<br><br>Acquaintance Details: <br>";
+        int j = 0;
+        Expression searchExpression = new Parser().constructParser(str);
+        for(int i = 0; i < a.get(0).size(); i++){
+            //if(a.get(0).get(i).getName().matches(str)){
+            if(searchExpression.interpret(a.get(0).get(i).getName())){
+                j++;
+                PersonalFriends perF = (PersonalFriends)a.get(0).get(i);
+                if(j==1){
+                    s = s.concat("<br>I. Personal Friends<br>");
+                }
+                s = s.concat(j + ". Name: " + perF.getName() + "<br>");
+                s = s.concat("Mobile No: " + perF.getMobileNo() + "<br>");
+                s = s.concat("Email: " + perF.getEmail() + "<br>");
+                s = s.concat("Specific events: " + perF.getEvents() + "<br>");
+                s = s.concat("First Acquaintance context: " + perF.getAContext() + "<br>");
+                s = s.concat("First Acquaintance date: " + perF.getADate() + "<br>");
+            }
+        }
+        j = 0;
+        for(int i = 0; i < a.get(1).size(); i++){
+            //if(a.get(1).get(i).getName().matches(str)){
+            if(searchExpression.interpret(a.get(1).get(i).getName())){
+                j++;
+                Relatives rel = (Relatives)a.get(1).get(i);
+                if(j==1){
+                    s = s.concat("<br>II. Relatives<br>");
+                }
+                s = s.concat(j + ". Name: " + rel.getName() + "<br>");
+                s = s.concat("Mobile No: " + rel.getMobileNo() + "<br>");
+                s = s.concat("Email: " + rel.getEmail() + "<br>");
+                s = s.concat("Relatives Birthday: " + rel.getBDate() + "<br>");
+                s = s.concat("Last met date: " + rel.getLDate() + "<br>");
+            }
+        }
+        j = 0;
+        for(int i = 0; i < a.get(2).size(); i++){
+            //if(a.get(2).get(i).getName().matches(str)){
+            if(searchExpression.interpret(a.get(2).get(i).getName())){
+                j++;
+                ProfessionalFriends proF = (ProfessionalFriends)a.get(2).get(i);
+                if(j==1){
+                    s = s.concat("<br>III. Professional Friends<br>");
+                }
+                s = s.concat(j + ". Name: " + proF.getName() + "<br>");
+                s = s.concat("Mobile No: " + proF.getMobileNo() + "<br>");
+                s = s.concat("Email: " + proF.getEmail() + "<br>");
+                s = s.concat("Common Interests: " + proF.getCommonInterests() + "<br>");
+            }
+        }
+        j = 0;
+        for(int i = 0; i < a.get(3).size(); i++){
+            //if(a.get(3).get(i).getName().matches(str)){
+            if(searchExpression.interpret(a.get(3).get(i).getName())){
+                j++;
+                CasualAcquaintances ca = (CasualAcquaintances)a.get(3).get(i);
+                if(j==1){
+                    s = s.concat("<br>IV. Casual Acquaintances<br>");
+                }
+                s = s.concat(j + ". Name: " + ca.getName() + "<br>");
+                s = s.concat("Mobile No: " + ca.getMobileNo() + "<br>");
+                s = s.concat("Email: " + ca.getEmail() + "<br>");
+                s = s.concat("First met location & time: " + ca.getWhenWhere() + "<br>");
+                s = s.concat("First met circumstances: " + ca.getCircumstances() + "<br>");
+                s = s.concat("Other useful information: " + ca.getOtherInfo() + "<br>");
+            }
+        }
+        if(s.matches("<html> <b>Search results:</b><br>Found!<br><br>Acquaintance Details: <br>")){
+            s  = "<html>No result found</html>";
+        }
+        else{
+            s = s.concat("</html>");
+        }
+        details.setText(s);
     }
 
     public Acquaintances getAcquaintance(){
