+++ src/contactmanagementsoftware/command/SaveCommand.java
@@ -11,8 +11,9 @@ public class SaveCommand implements Command {
     }
 
     @Override
-    public void execute(ActionEvent evt) {
+    public boolean execute(ActionEvent evt) {
         contactReceiver.saveContacts();
+        return true;
     }
 
     @Override
